requirements.txt
boto3
colorama
numpy
pandas
Pillow
PyYAML
tqdm